+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add version.py.
 * Viewer module: new registration matrix viewer.
 * Viewer module: new metadata viewer.
+* Add a minimal documentation for all modules.
 
 ### Changed